move application page into whitelist folder
